requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-redis==4.12.1
django-rest-swagger==2.1.2
djangorestframework==3.11.0
jsonpath==0.82
Markdown==3.2.2
mysqlclient==1.4.6
PyMySQL==0.8.1
python-docx==0.8.6
requests==2.18.4
python-crontab==2.8.1
python-dateutil==2.8.1
